+++ examples/android-room/src/main/java/io/sentry/android/roomsample/data/TracksDao.kt
@@ -14,7 +14,10 @@ abstract class TracksDao {
     @Query("SELECT * FROM track ORDER BY TrackId DESC")
     abstract fun all(): Flow<List<Track>>
 
-    @Query("SELECT * FROM Track WHERE AlbumId = (SELECT AlbumId FROM Album WHERE ArtistId = (SELECT ArtistId from Artist WHERE Name = :bandName))")
+    @Query(
+        "SELECT * FROM Track WHERE AlbumId = (SELECT AlbumId FROM Album WHERE ArtistId = " +
+            "(SELECT ArtistId from Artist WHERE Name = :bandName))"
+    )
     abstract suspend fun allByArtist(bandName: String): List<Track>
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)